+++ R/tar_config_set.R
@@ -200,10 +200,7 @@ tar_config_set <- function(
   tar_config_assert_store(store)
   tar_config_assert_use_crew(use_crew)
   tar_config_assert_workers(workers)
-  yaml <- tar_config_read_yaml(config)
-  if (!tar_config_is_multi_project(yaml, config)) {
-    yaml <- tar_config_convert_multi_project(yaml, config)
-  }
+  yaml <- tar_config_yaml(config = config)
   yaml[[project]]$inherits <- inherits %|||% yaml[[project]]$inherits
   yaml[[project]]$garbage_collection <- garbage_collection %|||%
     yaml[[project]]$garbage_collection